full_pipeline_demo/requirements.txt
# Full Pipeline Demo - Requirements
# ==================================
# Install: pip install -r requirements.txt

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
polars>=0.20.0

# Snowflake
snowflake-connector-python[pandas]>=3.0.0

# Machine Learning
xgboost>=1.7.0
scikit-learn>=1.3.0

# MLflow (optional - for experiment tracking)
mlflow>=2.0.0

# API & Web
requests>=2.28.0
fastapi>=0.100.0
uvicorn>=0.22.0

# dbt (for transformations)
dbt-snowflake>=1.6.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0
